+++ back/src/main/java/com/back/BackApplication.java
@@ -1,0 +1,13 @@
+package com.back;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BackApplication.class, args);
+	}
+
+}
