+++ back/src/main/java/com/back/domain/member/member/dto/LoginRequest.java
@@ -1,0 +1,9 @@
+package com.back.domain.member.member.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginRequest {
+    private String email;
+    private String password;
+}
