+++ back/src/main/java/com/back/domain/member/member/dto/SignupRequest.java
@@ -1,0 +1,12 @@
+package com.back.domain.member.member.dto;
+
+import com.back.domain.member.member.entity.Member;
+import lombok.Data;
+
+@Data
+public class SignupRequest {
+    private String email;
+    private String password;
+    private String name;
+    private Member.Role role;
+}
