+++ back/src/main/java/com/back/domain/member/member/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.back.domain.member.member.repository;
+
+import com.back.domain.member.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Integer> {
+    Optional<Member> findByEmail(String email);
+
+}
