+++ back/src/main/java/com/back/global/app/AppConfig.java
@@ -1,0 +1,38 @@
+package com.back.global.app;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.core.env.Environment;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class AppConfig {
+
+    private static Environment environment;
+
+    @Autowired
+    public void setEnvironment(Environment environment) {
+        AppConfig.environment = environment;
+    }
+
+    public static boolean isDev() {
+        return environment.matchesProfiles("dev");
+    }
+
+    public static boolean isTest() {
+        return environment.matchesProfiles("test");
+    }
+
+    public static boolean isProd() {
+        return environment.matchesProfiles("prod");
+    }
+
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
