+++ back/src/main/java/com/back/global/globalExceptionHandler/GlobalExceptionHandler.java
@@ -1,0 +1,122 @@
+package com.back.global.globalExceptionHandler;
+
+import com.back.global.exception.ServiceException;
+import com.back.global.rsData.RsData;
+import jakarta.validation.ConstraintViolationException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.MissingRequestHeaderException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.Comparator;
+import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
+
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+
+@RestControllerAdvice
+@RequiredArgsConstructor
+public class GlobalExceptionHandler {
+    @ExceptionHandler(NoSuchElementException.class)
+    public ResponseEntity<RsData<Void>> handle(NoSuchElementException ex) {
+        return new ResponseEntity<>(
+                new RsData<>(
+                        "404-1",
+                        "해당 데이터가 존재하지 않습니다."
+                ),
+                NOT_FOUND
+        );
+    }
+
+
+    @ExceptionHandler(ConstraintViolationException.class)
+    public ResponseEntity<RsData<Void>> handle(ConstraintViolationException ex) {
+        String message = ex.getConstraintViolations()
+                .stream()
+                .map(violation -> {
+                    String field = violation.getPropertyPath().toString().split("\\.", 2)[1];
+                    String[] messageTemplateBits = violation.getMessageTemplate()
+                            .split("\\.");
+                    String code = messageTemplateBits[messageTemplateBits.length - 2];
+                    String _message = violation.getMessage();
+
+                    return "%s-%s-%s".formatted(field, code, _message);
+                })
+                .sorted(Comparator.comparing(String::toString))
+                .collect(Collectors.joining("\n"));
+
+        return new ResponseEntity<>(
+                new RsData<>(
+                        "400-1",
+                        message
+                ),
+                BAD_REQUEST
+        );
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<RsData<Void>> handle(MethodArgumentNotValidException ex) {
+        String message = ex.getBindingResult()
+                .getAllErrors()
+                .stream()
+                .filter(error -> error instanceof FieldError)
+                .map(error -> (FieldError) error)
+                .map(error -> error.getField() + "-" + error.getCode() + "-" + error.getDefaultMessage())
+                .sorted(Comparator.comparing(String::toString))
+                .collect(Collectors.joining("\n"));
+
+        return new ResponseEntity<>(
+                new RsData<>(
+                        "400-1",
+                        message
+                ),
+                BAD_REQUEST
+        );
+    }
+
+    @ExceptionHandler(HttpMessageNotReadableException.class)
+    public ResponseEntity<RsData<Void>> handle(HttpMessageNotReadableException ex) {
+        return new ResponseEntity<>(
+                new RsData<>(
+                        "400-1",
+                        "요청 본문이 올바르지 않습니다."
+                ),
+                BAD_REQUEST
+        );
+    }
+
+    @ExceptionHandler(MissingRequestHeaderException.class)
+    public ResponseEntity<RsData<Void>> handle(MissingRequestHeaderException ex) {
+        return new ResponseEntity<>(
+                new RsData<>(
+                        "400-1",
+                        "%s-%s-%s".formatted(
+                                ex.getHeaderName(),
+                                "NotBlank",
+                                ex.getLocalizedMessage()
+                        )
+                ),
+                BAD_REQUEST
+        );
+    }
+
+    @ExceptionHandler(ServiceException.class)
+    @ResponseStatus(BAD_REQUEST)
+    public ResponseEntity<RsData<Void>> handle(ServiceException ex) {
+        RsData<Void> rsData = ex.getRsData();
+
+        return new ResponseEntity<>(
+                rsData,
+                ResponseEntity
+                        .status(rsData.statusCode())
+                        .build()
+                        .getStatusCode()
+        );
+    }
+}
