+++ back/src/main/java/com/back/global/rsData/RsData.java
@@ -1,0 +1,18 @@
+package com.back.global.rsData;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
+public record RsData<T>(
+        String resultCode,
+        @JsonIgnore int statusCode,
+        String msg,
+        T data
+) {
+    public RsData(String resultCode, String msg) {
+        this(resultCode, msg, null);
+    }
+
+    public RsData(String resultCode, String msg, T data) {
+        this(resultCode, Integer.parseInt(resultCode.split("-", 2)[0]), msg, data);
+    }
+}
