+++ back/src/main/java/com/back/global/security/SecurityUser.java
@@ -1,0 +1,28 @@
+package com.back.global.security;
+
+
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+
+import java.util.Collection;
+
+public class SecurityUser extends User {
+    @Getter
+    private int id;
+
+    @Getter
+    private String name;
+
+    public SecurityUser(
+            int id,
+            String username,
+            String password,
+            String name,
+            Collection<? extends GrantedAuthority> authorities
+    ) {
+        super(username, password != null ? password : "", authorities);
+        this.id = id;
+        this.name = name;
+    }
+}
