+++ back/src/main/java/com/back/standard/util/Ut.java
@@ -1,0 +1,88 @@
+package com.back.standard.util;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ClaimsBuilder;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+
+import javax.crypto.SecretKey;
+import java.security.Key;
+import java.util.Date;
+import java.util.Map;
+
+public class Ut {
+    public static class jwt {
+        public static String toString(String secret, int expireSeconds, Map<String, Object> body) {
+            ClaimsBuilder claimsBuilder = Jwts.claims();
+
+            for (Map.Entry<String, Object> entry : body.entrySet()) {
+                claimsBuilder.add(entry.getKey(), entry.getValue());
+            }
+
+            Claims claims = claimsBuilder.build();
+
+            Date issuedAt = new Date();
+            Date expiration = new Date(issuedAt.getTime() + 1000L * expireSeconds);
+
+            Key secretKey = Keys.hmacShaKeyFor(secret.getBytes());
+
+            String jwt = Jwts.builder()
+                    .claims(claims)
+                    .issuedAt(issuedAt)
+                    .expiration(expiration)
+                    .signWith(secretKey)
+                    .compact();
+
+            return jwt;
+        }
+
+        public static boolean isValid(String secret, String jwtStr) {
+            SecretKey secretKey = Keys.hmacShaKeyFor(secret.getBytes());
+
+            try {
+                Jwts
+                        .parser()
+                        .verifyWith(secretKey)
+                        .build()
+                        .parse(jwtStr);
+            } catch (Exception e) {
+                return false;
+            }
+
+            return true;
+        }
+
+        public static Map<String, Object> payload(String secret, String jwtStr) {
+            SecretKey secretKey = Keys.hmacShaKeyFor(secret.getBytes());
+
+            try {
+                return (Map<String, Object>) Jwts
+                        .parser()
+                        .verifyWith(secretKey)
+                        .build()
+                        .parse(jwtStr)
+                        .getPayload();
+            } catch (Exception e) {
+                return null;
+            }
+        }
+    }
+
+    public static class json {
+        public static ObjectMapper objectMapper = new ObjectMapper();
+
+        public static String toString(Object object) {
+            return toString(object, null);
+        }
+
+        public static String toString(Object object, String defaultValue) {
+            try {
+                return objectMapper.writeValueAsString(object);
+            } catch (Exception e) {
+                return defaultValue;
+            }
+        }
+    }
+
+}
