--- back/src/main/java/com/ll/back/BackApplication.java
@@ -1,13 +1,0 @@
-package com.ll.back;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackApplication.class, args);
-	}
-
-}
