--- back/src/main/java/com/ll/back/domain/member/member/dto/LoginRequest.java
@@ -1,9 +1,0 @@
-package com.ll.back.domain.member.member.dto;
-
-import lombok.Data;
-
-@Data
-public class LoginRequest {
-    private String email;
-    private String password;
-}
