--- back/src/main/java/com/ll/back/domain/member/member/entity/Member.java
@@ -1,44 +1,0 @@
-package com.ll.back.domain.member.member.entity;
-
-import com.ll.back.global.jpa.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Member extends BaseEntity {
-    @Column(unique = true, nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Role role;
-
-    public enum Role {
-        MENTOR, MENTEE, ADMIN
-    }
-
-    public Member(String email, String password, String name, Role role) {
-        this.email = email;
-        this.password = password;
-        this.name = name;
-        this.role = role;
-    }
-
-    public Member(int id, String email, String name) {
-        this.setId(id);
-        this.email = email;
-        this.name = name;
-    }
-}
