--- back/src/main/java/com/ll/back/domain/member/member/repository/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.ll.back.domain.member.member.repository;
-
-import com.ll.back.domain.member.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Integer> {
-    Optional<Member> findByEmail(String email);
-
-}
