--- back/src/main/java/com/ll/back/global/exception/ServiceException.java
@@ -1,18 +1,0 @@
-package com.ll.back.global.exception;
-
-import com.ll.back.global.rsData.RsData;
-
-public class ServiceException extends RuntimeException {
-    private final String resultCode;
-    private final String msg;
-
-    public ServiceException(String resultCode, String msg) {
-        super(resultCode + " : " + msg);
-        this.resultCode = resultCode;
-        this.msg = msg;
-    }
-
-    public RsData<Void> getRsData() {
-        return new RsData<>(resultCode, msg, null);
-    }
-}
